+++ packages/gamut/src/Disclosure/helpers.tsx
@@ -1,4 +1,4 @@
-import { FillButton, StrokeButton, TextButton } from "../Button";
+import { FillButton, StrokeButton, TextButton } from '../Button';
 
 const verticalSpacingMap = {
   normal: 16,
@@ -24,23 +24,32 @@ export const determineHorizontalSpacing = (
   return horizontalSpacingMap[spacing];
 };
 
-export const renderButton = (buttonName: string, ctaText: string , ctaCallback: ()=>void) => {
-  if(buttonName.match(/fill/i)) {
-    return(<FillButton
-      alignSelf="flex-end"
-      pt={8}
-      onClick={ctaCallback ? () => ctaCallback() : undefined}
-    >
-      {ctaText}
-    </FillButton>)
-  } if (buttonName.match(/stroke/i)) {
-    return(<StrokeButton
-      alignSelf="flex-end"
-      pt={8}
-      onClick={ctaCallback ? () => ctaCallback() : undefined}
-    >
-      {ctaText}
-    </StrokeButton>)
+export const renderButton = (
+  buttonName: string,
+  ctaText: string,
+  ctaCallback: () => void
+) => {
+  if (buttonName.match(/fill/i)) {
+    return (
+      <FillButton
+        alignSelf="flex-end"
+        pt={8}
+        onClick={ctaCallback ? () => ctaCallback() : undefined}
+      >
+        {ctaText}
+      </FillButton>
+    );
+  }
+  if (buttonName.match(/stroke/i)) {
+    return (
+      <StrokeButton
+        alignSelf="flex-end"
+        pt={8}
+        onClick={ctaCallback ? () => ctaCallback() : undefined}
+      >
+        {ctaText}
+      </StrokeButton>
+    );
   }
   return (
     <TextButton
@@ -50,6 +59,5 @@ export const renderButton = (buttonName: string, ctaText: string , ctaCallback: 
     >
       {ctaText}
     </TextButton>
-  )
-}
-
+  );
+};
