+++ packages/gamut/src/Disclosure/index.tsx
@@ -1,32 +1,14 @@
 import { variant } from '@codecademy/gamut-styles';
 import styled from '@emotion/styled';
-import { AnimatePresence, motion } from 'framer-motion';
+import { AnimatePresence } from 'framer-motion';
 import { useState } from 'react';
 import * as React from 'react';
 
-import { WithChildrenProp } from '..';
+import { ExpandInCollapseOut } from '../Animation/ExpandInCollapseOut';
 import { FlexBox } from '../Box';
-import { DisclosureArea } from '../DisclosureArea';
-import { DisclosureButton } from '../DisclosureButton';
+import { DisclosureBody } from './DisclosureBody';
+import { DisclosureButton } from './DisclosureButton';
 import { DisclosureProps } from './types';
-
-const ExpandInCollapseOut: React.FC<WithChildrenProp> = ({ children }) => {
-  return (
-    <motion.div
-      initial="collapsed"
-      exit="collapsed"
-      animate="expanded"
-      style={{ overflow: 'hidden' }}
-      variants={{
-        expanded: { height: 'auto' },
-        collapsed: { height: 0 },
-      }}
-      transition={{ duration: 0.2, ease: 'easeInOut' }}
-    >
-      {children}
-    </motion.div>
-  );
-};
 
 const StyledFlexBox = styled(FlexBox)(
   variant({
@@ -35,6 +17,7 @@ const StyledFlexBox = styled(FlexBox)(
       default: {
         bg: 'background-current',
         border: 1,
+        maxHeight: 'fit-content',
       },
       block: {
         bg: 'background',
@@ -44,17 +27,15 @@ const StyledFlexBox = styled(FlexBox)(
   })
 );
 
-// Appearances can be made optional and given a reasonable default value
 export const Disclosure: React.FC<DisclosureProps> = ({
   disabled = false,
   header,
   variant,
   headingLevel = 'h3',
-  initiallyExpanded = true,
-  // Remove these defaults later (maybe keep spacing)
-  overline = 'overline is optional',
+  initiallyExpanded,
   spacing = 'normal',
-  subheader = 'subheader is optional',
+  overline,
+  subheader,
   body,
   withBackground,
   ctaText,
@@ -74,10 +55,11 @@ export const Disclosure: React.FC<DisclosureProps> = ({
         setIsExpanded={setIsExpanded}
         disabled={disabled}
       />
-      <AnimatePresence>
-        {isExpanded && (
+
+      {isExpanded && (
+        <AnimatePresence>
           <ExpandInCollapseOut>
-            <DisclosureArea
+            <DisclosureBody
               body={body}
               withBackground={withBackground}
               ctaText={ctaText}
@@ -85,8 +67,8 @@ export const Disclosure: React.FC<DisclosureProps> = ({
               ctaCallback={ctaCallback}
             />
           </ExpandInCollapseOut>
-        )}
-      </AnimatePresence>
+        </AnimatePresence>
+      )}
     </StyledFlexBox>
   );
 };
