--- packages/styleguide/stories/Molecules/Disclosure.stories.mdx
@@ -1,30 +1,0 @@
-import { Disclosure } from '@codecademy/gamut/src/Disclosure';
-import title from '@codecademy/macros/lib/title.macro';
-import { Meta, Story } from '@storybook/addon-docs/blocks';
-
-// TO ASK: why does this need `.../src/Disclosure` when other imports don't?
-
-<Meta
-  title={title}
-  component={Disclosure}
-  parameters={{
-    subtitle: `
-A molecule consisting of a “top” section and an expandable body, such as a “hint” section in the LE or a list of department jobs on a careers page.
-`,
-    source: 'gamut',
-    status: 'current',
-  }}
-  args={{
-    header:
-      'this is the header. Lorem ipsum dolor sit amet, consectetur adipiscing elit.',
-    body:
-      'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur.',
-    display: 'background',
-  }}
-/>
-
-## This is a sample
-
-Edit attributes and see what’s possible to do with this component or get the exact output you want.
-
-<Story name="Disclosure new">{(args) => <Disclosure {...args} />}</Story>
